pyproject.toml
[project]
name = "data-time-machine"
version = "0.1.1"
description = "A git-like state management system for data pipelines"
readme = "README.md"
authors = [
    {name = "Azmat Siddique", email = "azmat.siddique.98@gmail.com"},
]
dependencies = [
    "click>=8.1.7",
    "pydantic>=2.5.3",
]

[project.optional-dependencies]
test = ["pytest>=7.4.0"]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "jinja2>=3.1.0",
]
cloud = [
    "boto3>=1.28.0",
    "google-cloud-storage>=2.10.0",
    "azure-storage-blob>=12.17.0",
]
all = [
    "data-time-machine[web,cloud]",
]

requires-python = ">=3.10"

[project.urls]
Homepage = "https://github.com/azmatsiddique/data-time-machine"
Repository = "https://github.com/azmatsiddique/data-time-machine"
Issues = "https://github.com/azmatsiddique/data-time-machine/issues"

[project.scripts]
dtm = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
